Bug Fix for deleting Ingest Sheets
